+++ examples/6lbr-demo/apps/coap/coap-common.h
@@ -84,10 +84,10 @@ extern unsigned long coap_batch_basetime;
 /*---------------------------------------------------------------------------*/
 
 int
-coap_strtoul(char const *data, char const *max, unsigned long *value);
+coap_strtoul(char const *data, char const *max, uint32_t *value);
 
 int
-coap_strtofix(char const *data, char const *max, unsigned long *value, int precision);
+coap_strtofix(char const *data, char const *max, uint32_t *value, int precision);
 
 #define REST_PARSE_EMPTY(payload, len, actuator_set) {\
   if(len==0) {\
@@ -106,7 +106,7 @@ coap_strtofix(char const *data, char const *max, unsigned long *value, int preci
 }
 
 #define REST_PARSE_ONE_UINT(buffer, max, data) { \
-  unsigned long value; \
+  uint32_t value; \
   if (coap_strtoul((char const *)buffer, max, &value)) { \
     data = value; \
     success = 1; \
@@ -114,7 +114,7 @@ coap_strtofix(char const *data, char const *max, unsigned long *value, int preci
 }
 
 #define REST_PARSE_ONE_DECIMAL(buffer, max, data) { \
-  unsigned long value; \
+  uint32_t value; \
   if (coap_strtofix((char const *)buffer, max, &value, 1)) { \
     data = value; \
     success = 1; \
@@ -171,12 +171,12 @@ full_resource_config_handler(coap_full_resource_t *resource_info, void* request,
 #endif
 
 #define REST_FULL_RESOURCE_GET_HANDLER(resource_name, format, parser) \
-  int resource_##resource_name##_format_value(char *buffer, unsigned long data) {\
+  int resource_##resource_name##_format_value(char *buffer, uint32_t data) {\
     int len; \
     format(resource_name, data); \
     return len; \
   } \
- int resource_##resource_name##_parse_value(char const *buffer, char const *max, unsigned long *data) {\
+ int resource_##resource_name##_parse_value(char const *buffer, char const *max, uint32_t *data) {\
    int success = 0; \
    parser(buffer, max, *data); \
    return success; \
@@ -264,8 +264,8 @@ full_resource_config_handler(coap_full_resource_t *resource_info, void* request,
 #define REST_FULL_RESOURCE(resource_name, resource_period, resource_if, resource_type, resource_format, resource_parse, resource_id, resource_value) \
   RESOURCE_DECL(resource_name); \
   extern void update_resource_##resource_name##_value(coap_resource_data_t *data); \
-  extern int resource_##resource_name##_format_value(char * buffer, unsigned long data); \
-  extern int resource_##resource_name##_parse_value(char const *buffer, char const *max, unsigned long *data); \
+  extern int resource_##resource_name##_format_value(char * buffer, uint32_t data); \
+  extern int resource_##resource_name##_parse_value(char const *buffer, char const *max, uint32_t *data); \
   coap_full_resource_t resource_##resource_name##_info = { \
       .coap_resource = &resource_##resource_name, \
       .trigger = { .flags = resource_period != 0 ? COAP_BINDING_FLAGS_PMIN_VALID : 0, .pmin = resource_period }, \
